improve parseXML to use callback
